backend/requirements.txt
altgraph==0.10.2
aniso8601==1.3.0
appdirs==1.4.3
astroid==1.5.3
attrs==17.2.0
Automat==0.6.0
backports.functools-lru-cache==1.4
bdist-mpkg==0.5.0
beautifulsoup4==4.5.3
bonjour-py==0.3
click==6.7
configparser==3.5.0
constantly==15.1.0
dash==0.17.5
dash-core-components==0.5.0
dash-html-components==0.6.1
dash-renderer==0.7.3
dash.ly==0.17.3
decorator==4.0.11
Django==1.10.5
django-bootstrap3==8.1.0
django-cors-headers==2.1.0
django-dotenv==1.4.1
django-extensions==1.7.9
django-filter==1.0.1
django-pipeline==1.6.13
django-pipeline-browserify==0.6.1
django-twitter-bootstrap==3.3.0
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
enum34==1.1.6
functools32==3.2.3.post2
futures==3.1.1
incremental==17.5.0
ipython-genutils==0.2.0
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter-core==4.3.0
lazy-object-proxy==1.3.1
macholib==1.5.1
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==1.3.1
mccabe==0.6.1
modulegraph==0.10.4
nbformat==4.3.0
numpy==1.8.0rc1
olefile==0.44
ordereddict==1.1
packaging==16.8
pandas==0.20.2
Pillow==4.2.1
plotly==2.0.10
psycopg2==2.6.2
py2app==0.7.3
Pygments==2.2.0
PyJWT==1.4.2
pylint==1.7.1
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
pysolr==3.5.0
python-dateutil==1.5
python-dotenv==0.6.5
PyYAML==3.12
requests==2.13.0
scipy==0.13.0b1
singledispatch==3.4.0.3
six==1.4.1
SQLAlchemy==1.1.15
traitlets==4.3.2
turtle==0.0.2
Twisted==17.1.0
vboxapi==1.0
virtualenv==15.0.3
watson-developer-cloud==0.21.0
Werkzeug==0.12.2
Willow==0.4
wrapt==1.10.10
xattr==0.6.4
zope.interface==4.1.1
